Skin cleanup
The theme was working in some situation and broken in others and I couldn’t find the cause.

After trials and error, the way the 'renderEditButton' method was built would break
because of the methods were showing broken PHP tags and that would create a race-condition.

This is to rewrite the template file so that only WebPlatformTemplate::execute() would be like that.